+++ src/dsp/fsk_demod.h
@@ -8,7 +8,7 @@
 
 typedef struct fsk_demod_t fsk_demod;
 
-int fsk_demod_create(uint32_t sampling_freq, uint32_t baud_rate, int32_t deviation, uint8_t decimation, uint32_t transition_width, bool use_dc_block, uint32_t max_input_buffer_length, fsk_demod **demod);
+int fsk_demod_create(uint64_t sampling_freq, uint32_t baud_rate, int64_t deviation, uint8_t decimation, uint32_t transition_width, bool use_dc_block, uint32_t max_input_buffer_length, fsk_demod **demod);
 
 void fsk_demod_process(const float complex *input, size_t input_len, int8_t **output, size_t *output_len, fsk_demod *demod);
