+++ src/dsp/lpf_taps.h
@@ -3,6 +3,6 @@
 
 #include <stdint.h>
 
-int create_low_pass_filter(float gain, uint32_t sampling_freq, uint32_t cutoff_freq, uint32_t transition_width, float **taps, size_t *len);
+int create_low_pass_filter(float gain, uint64_t sampling_freq, uint64_t cutoff_freq, uint32_t transition_width, float **taps, size_t *len);
 
 #endif /* SRC_LPF_H_ */
